+++ src/fmt/mod.rs
@@ -37,15 +37,11 @@ use std::fmt::Display;
 
 use log::Record;
 
-mod termcolor;
-mod atty;
+mod writer;
 mod humantime;
 
-use self::termcolor::{Buffer, BufferWriter};
-use self::atty::{is_stdout, is_stderr};
-
 pub use self::humantime::pub_use_in_super::*;
-pub use self::termcolor::pub_use_in_super::*;
+pub use self::writer::pub_use_in_super::*;
 
 pub(super) mod pub_use_in_super {
     pub use super::{Target, WriteStyle, Formatter};
@@ -210,7 +206,7 @@ impl<'a> DefaultFormat<'a> {
     fn finish_header(&mut self) -> io::Result<()> {
         if self.written_header_value {
             let close_brace = self.subtle_style("]");
-            write!(self.buf, "{}", close_brace)
+            write!(self.buf, "{} ", close_brace)
         } else {
             Ok(())
         }
@@ -246,126 +242,10 @@ pub struct Formatter {
     write_style: WriteStyle,
 }
 
-/// Log target, either `stdout` or `stderr`.
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub enum Target {
-    /// Logs will be sent to standard output.
-    Stdout,
-    /// Logs will be sent to standard error.
-    Stderr,
-}
-
-impl Default for Target {
-    fn default() -> Self {
-        Target::Stderr
-    }
-}
-
-/// Whether or not to print styles to the target.
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub enum WriteStyle {
-    /// Try to print styles, but don't force the issue.
-    Auto,
-    /// Try very hard to print styles.
-    Always,
-    /// Never print styles.
-    Never,
-}
-
-impl Default for WriteStyle {
-    fn default() -> Self {
-        WriteStyle::Auto
-    }
-}
-
-/// A terminal target with color awareness.
-pub(crate) struct Writer {
-    inner: BufferWriter,
-    write_style: WriteStyle,
-}
-
-impl Writer {
-    pub(crate) fn write_style(&self) -> WriteStyle {
-        self.write_style
-    }
-}
-
-/// A builder for a terminal writer.
-///
-/// The target and style choice can be configured before building.
-pub(crate) struct Builder {
-    target: Target,
-    write_style: WriteStyle,
-}
-
-impl Builder {
-    /// Initialize the writer builder with defaults.
-    pub fn new() -> Self {
-        Builder {
-            target: Default::default(),
-            write_style: Default::default(),
-        }
-    }
-
-    /// Set the target to write to.
-    pub fn target(&mut self, target: Target) -> &mut Self {
-        self.target = target;
-        self
-    }
-
-    /// Parses a style choice string.
-    ///
-    /// See the [Disabling colors] section for more details.
-    ///
-    /// [Disabling colors]: ../index.html#disabling-colors
-    pub fn parse(&mut self, write_style: &str) -> &mut Self {
-        self.write_style(parse_write_style(write_style))
-    }
-
-    /// Whether or not to print style characters when writing.
-    pub fn write_style(&mut self, write_style: WriteStyle) -> &mut Self {
-        self.write_style = write_style;
-        self
-    }
-
-    /// Build a terminal writer.
-    pub fn build(&mut self) -> Writer {
-        let color_choice = match self.write_style {
-            WriteStyle::Auto => {
-                if match self.target {
-                    Target::Stderr => is_stderr(),
-                    Target::Stdout => is_stdout(),
-                } {
-                    WriteStyle::Auto
-                } else {
-                    WriteStyle::Never
-                }
-            },
-            color_choice => color_choice,
-        };
-
-        let writer = match self.target {
-            Target::Stderr => BufferWriter::stderr(color_choice),
-            Target::Stdout => BufferWriter::stdout(color_choice),
-        };
-
-        Writer {
-            inner: writer,
-            write_style: self.write_style,
-        }
-    }
-}
-
-impl Default for Builder {
-    fn default() -> Self {
-        Builder::new()
-    }
-}
-
 impl Formatter {
     pub(crate) fn new(writer: &Writer) -> Self {
         Formatter {
-            buf: Rc::new(RefCell::new(writer.inner.buffer())),
+            buf: Rc::new(RefCell::new(writer.buffer())),
             write_style: writer.write_style(),
         }
     }
@@ -375,7 +255,7 @@ impl Formatter {
     }
 
     pub(crate) fn print(&self, writer: &Writer) -> io::Result<()> {
-        writer.inner.print(&self.buf.borrow())
+        writer.print(&self.buf.borrow())
     }
 
     pub(crate) fn clear(&mut self) {
@@ -402,55 +282,5 @@ impl fmt::Debug for Writer {
 impl fmt::Debug for Formatter {
     fn fmt(&self, f: &mut fmt::Formatter)->fmt::Result {
         f.debug_struct("Formatter").finish()
-    }
-}
-
-impl fmt::Debug for Builder {
-    fn fmt(&self, f: &mut fmt::Formatter)->fmt::Result {
-        f.debug_struct("Logger")
-        .field("target", &self.target)
-        .field("write_style", &self.write_style)
-        .finish()
-    }
-}
-
-fn parse_write_style(spec: &str) -> WriteStyle {
-    match spec {
-        "auto" => WriteStyle::Auto,
-        "always" => WriteStyle::Always,
-        "never" => WriteStyle::Never,
-        _ => Default::default(),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parse_write_style_valid() {
-        let inputs = vec![
-            ("auto", WriteStyle::Auto),
-            ("always", WriteStyle::Always),
-            ("never", WriteStyle::Never),
-        ];
-
-        for (input, expected) in inputs {
-            assert_eq!(expected, parse_write_style(input));
-        }
-    }
-
-    #[test]
-    fn parse_write_style_invalid() {
-        let inputs = vec![
-            "",
-            "true",
-            "false",
-            "NEVER!!"
-        ];
-
-        for input in inputs {
-            assert_eq!(WriteStyle::Auto, parse_write_style(input));
-        }
     }
 }
