+++ src/fmt/writer/termcolor/shim_impl.rs
@@ -2,7 +2,7 @@ use std::io::{self, Write};
 
 use fmt::{WriteStyle, Target};
 
-pub(in ::fmt) mod pub_use_in_super {
+pub(in ::fmt::writer) mod pub_use_in_super {
     
 }
 
@@ -13,23 +13,23 @@ pub(in ::fmt) struct BufferWriter {
 pub(in ::fmt) struct Buffer(Vec<u8>);
 
 impl BufferWriter {
-    pub(in ::fmt) fn stderr(_: WriteStyle) -> Self {
+    pub(in ::fmt::writer) fn stderr(_: WriteStyle) -> Self {
         BufferWriter {
             target: Target::Stderr,
         }
     }
 
-    pub(in ::fmt) fn stdout(_: WriteStyle) -> Self {
+    pub(in ::fmt::writer) fn stdout(_: WriteStyle) -> Self {
         BufferWriter {
             target: Target::Stdout,
         }
     }
 
-    pub(in ::fmt) fn buffer(&self) -> Buffer {
+    pub(in ::fmt::writer) fn buffer(&self) -> Buffer {
         Buffer(Vec::new())
     }
 
-    pub(in ::fmt) fn print(&self, buf: &Buffer) -> io::Result<()> {
+    pub(in ::fmt::writer) fn print(&self, buf: &Buffer) -> io::Result<()> {
         match self.target {
             Target::Stderr => {
                 let stderr = io::stderr();
